+++ alikafkastreaming-20260202/alibabacloud_alikafkastreaming20260202/models/_create_compute_instance_request.py
@@ -10,14 +10,12 @@ class CreateComputeInstanceRequest(DaraModel):
         paid_type: int = None,
         region_id: str = None,
         resource_group_id: str = None,
-        resource_type: str = None,
     ):
         # This parameter is required.
         self.paid_type = paid_type
         # This parameter is required.
         self.region_id = region_id
         self.resource_group_id = resource_group_id
-        self.resource_type = resource_type
 
     def validate(self):
         pass
@@ -36,9 +34,6 @@ class CreateComputeInstanceRequest(DaraModel):
         if self.resource_group_id is not None:
             result['ResourceGroupId'] = self.resource_group_id
 
-        if self.resource_type is not None:
-            result['ResourceType'] = self.resource_type
-
         return result
 
     def from_map(self, m: dict = None):
@@ -51,9 +46,6 @@ class CreateComputeInstanceRequest(DaraModel):
 
         if m.get('ResourceGroupId') is not None:
             self.resource_group_id = m.get('ResourceGroupId')
-
-        if m.get('ResourceType') is not None:
-            self.resource_type = m.get('ResourceType')
 
         return self