+++ alikafkastreaming-20260202/alibabacloud_alikafkastreaming20260202/models/_list_compute_instances_in_page_request.py
@@ -10,19 +10,17 @@ class ListComputeInstancesInPageRequest(DaraModel):
     def __init__(
         self,
         current_page: int = None,
-        instance_id: str = None,
         instance_ids: List[str] = None,
-        order_id: str = None,
         page_size: int = None,
         region_id: str = None,
+        resource_group_id: str = None,
     ):
         self.current_page = current_page
-        self.instance_id = instance_id
         self.instance_ids = instance_ids
-        self.order_id = order_id
         self.page_size = page_size
         # This parameter is required.
         self.region_id = region_id
+        self.resource_group_id = resource_group_id
 
     def validate(self):
         pass
@@ -35,20 +33,17 @@ class ListComputeInstancesInPageRequest(DaraModel):
         if self.current_page is not None:
             result['CurrentPage'] = self.current_page
 
-        if self.instance_id is not None:
-            result['InstanceId'] = self.instance_id
-
         if self.instance_ids is not None:
             result['InstanceIds'] = self.instance_ids
-
-        if self.order_id is not None:
-            result['OrderId'] = self.order_id
 
         if self.page_size is not None:
             result['PageSize'] = self.page_size
 
         if self.region_id is not None:
             result['RegionId'] = self.region_id
+
+        if self.resource_group_id is not None:
+            result['ResourceGroupId'] = self.resource_group_id
 
         return result
 
@@ -57,20 +52,17 @@ class ListComputeInstancesInPageRequest(DaraModel):
         if m.get('CurrentPage') is not None:
             self.current_page = m.get('CurrentPage')
 
-        if m.get('InstanceId') is not None:
-            self.instance_id = m.get('InstanceId')
-
         if m.get('InstanceIds') is not None:
             self.instance_ids = m.get('InstanceIds')
-
-        if m.get('OrderId') is not None:
-            self.order_id = m.get('OrderId')
 
         if m.get('PageSize') is not None:
             self.page_size = m.get('PageSize')
 
         if m.get('RegionId') is not None:
             self.region_id = m.get('RegionId')
+
+        if m.get('ResourceGroupId') is not None:
+            self.resource_group_id = m.get('ResourceGroupId')
 
         return self