+++ alikafkastreaming-20260202/alibabacloud_alikafkastreaming20260202/models/_stop_compute_job_request.py
@@ -7,12 +7,10 @@ from darabonba.model import DaraModel
 class StopComputeJobRequest(DaraModel):
     def __init__(
         self,
-        client_token: str = None,
         instance_id: str = None,
         job_name: str = None,
         region_id: str = None,
     ):
-        self.client_token = client_token
         # This parameter is required.
         self.instance_id = instance_id
         # This parameter is required.
@@ -28,9 +26,6 @@ class StopComputeJobRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
-        if self.client_token is not None:
-            result['ClientToken'] = self.client_token
-
         if self.instance_id is not None:
             result['InstanceId'] = self.instance_id
 
@@ -44,9 +39,6 @@ class StopComputeJobRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
-        if m.get('ClientToken') is not None:
-            self.client_token = m.get('ClientToken')
-
         if m.get('InstanceId') is not None:
             self.instance_id = m.get('InstanceId')