+++ alikafkastreaming-20260202/alibabacloud_alikafkastreaming20260202/models/_update_compute_job_request.py
@@ -7,14 +7,11 @@ from darabonba.model import DaraModel
 class UpdateComputeJobRequest(DaraModel):
     def __init__(
         self,
-        client_token: str = None,
         instance_id: str = None,
         job_name: str = None,
         region_id: str = None,
         remark: str = None,
-        upgrade_mode: str = None,
     ):
-        self.client_token = client_token
         # This parameter is required.
         self.instance_id = instance_id
         # This parameter is required.
@@ -22,7 +19,6 @@ class UpdateComputeJobRequest(DaraModel):
         # This parameter is required.
         self.region_id = region_id
         self.remark = remark
-        self.upgrade_mode = upgrade_mode
 
     def validate(self):
         pass
@@ -32,9 +28,6 @@ class UpdateComputeJobRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
-        if self.client_token is not None:
-            result['ClientToken'] = self.client_token
-
         if self.instance_id is not None:
             result['InstanceId'] = self.instance_id
 
@@ -47,16 +40,10 @@ class UpdateComputeJobRequest(DaraModel):
         if self.remark is not None:
             result['Remark'] = self.remark
 
-        if self.upgrade_mode is not None:
-            result['UpgradeMode'] = self.upgrade_mode
-
         return result
 
     def from_map(self, m: dict = None):
         m = m or dict()
-        if m.get('ClientToken') is not None:
-            self.client_token = m.get('ClientToken')
-
         if m.get('InstanceId') is not None:
             self.instance_id = m.get('InstanceId')
 
@@ -68,9 +55,6 @@ class UpdateComputeJobRequest(DaraModel):
 
         if m.get('Remark') is not None:
             self.remark = m.get('Remark')
-
-        if m.get('UpgradeMode') is not None:
-            self.upgrade_mode = m.get('UpgradeMode')
 
         return self